requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==5.1.3
pillow==11.0.0
psycopg2==2.9.10
qr==0.6.0
qrcode==8.0
redis==5.2.0
sqlparse==0.5.2
typing_extensions==4.12.2
gunicorn
